+++ src/main.tsx
@@ -1,11 +1,16 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
 import App from './App'
+import ErrorBoundary from './components/ErrorBoundary'
 import './i18n'
 import './index.css'
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
-    <App />
+    <ErrorBoundary
+      storagePrefixes={["agile-toolkit:"]}
+    >
+      <App />
+    </ErrorBoundary>
   </React.StrictMode>,
 )